requirements.txt.txt
streamlit==1.12.0
pillow==9.2.0